Splitting the Node Collector in 2 Parts
